+++ addon/components/flexberry-checkbox.js
@@ -2,23 +2,58 @@
   @module ember-flexberry
 */
 
-import Ember from 'ember';
-import FlexberryBaseComponent from './flexberry-base-component';
+import FlexberryBaseComponent from 'ember-flexberry/components/flexberry-base-component';
 
 /**
-  Checkbox component for Semantic UI.
+  Flexberry checkbox component with [Semantic UI checkbox](http://semantic-ui.com/modules/checkbox.html) style.
 
-  @example
-    ```handlebars
-    {{flexberry-checkbox checked=model.enabled label='Enabled'}}
-    ```
+  Usage:
+  templates/my-form.hbs
+  ```handlebars
+  {{flexberry-checkbox
+    value=model.flag
+    onChange=(action "onModelFlagChange")
+  }}
+  ```
+
+  controllers/my-form.js
+  ```javascript
+  actions: {
+    onModelFlagChange(e) {
+      console.log('Model flag changed: ', e.checked);
+    }
+  }
+  ```
 
   @class FlexberryCheckboxComponent
   @extends FlexberryBaseComponent
 */
 export default FlexberryBaseComponent.extend({
   /**
-    Current checked value.
+    Component's wrapping <div> CSS-classes names.
+
+    Any other CSS-classes can be added through component's 'class' property.
+    ```handlebars
+    {{flexberry-checkbox class="toggle" value=model.flag}}
+    ```
+
+    @property classNames
+    @type String[]
+    @default ['flexberry-checkbox', 'ui', 'checkbox']
+  */
+  classNames: ['flexberry-checkbox', 'ui', 'checkbox'],
+
+  /**
+    Components class names bindings.
+
+    @property classNameBindings
+    @type String[]
+    @default ['value:checked', 'readonly:read-only']
+  */
+  classNameBindings: ['value:checked', 'readonly:read-only'],
+
+  /**
+    Component's value (if true, then checkbox is checked).
 
     @property value
     @type Boolean
@@ -27,53 +62,52 @@ export default FlexberryBaseComponent.extend({
   value: false,
 
   /**
-    Classes for component wrapper.
-  */
-  class: '',
-
-  /**
-    Label for the checkbox.
+    Component's label.
 
     @property label
     @type String
-  */
-  label: undefined,
-
-  /**
-    DOM-element representing checkbox input.
-
-    @property checkboxInput
-    @type Object
     @default null
   */
-  checkboxInput: null,
+  label: null,
 
   /**
-    Overload wrapper tag name for disabling wrapper.
-    [More info](http://emberjs.com/api/classes/Ember.Component.html#property_tagName).
-
-    @property tagName
-    @type String
-    @default ''
-    @readOnly
+    Initializes DOM-related component's properties.
   */
-  tagName: '',
+  didInsertElement() {
+    this._super(...arguments);
 
-  /**
-    Path to component's settings in application configuration (JSON from ./config/environment.js).
-
-    @property appConfigSettingsPath
-    @type String
-    @default 'APP.components.flexberryCheckbox'
-  */
-  appConfigSettingsPath: 'APP.components.flexberryCheckbox',
-
-  /**
-    Checkbox value's observer.
-  */
-  _valueDidChange: Ember.observer('value', (self) => {
-    self.sendAction('onChange', {
-      checked: self.get('value'),
+    // Initialize Semantic UI checkbox.
+    this.$().checkbox({
+      onChecked: () => {
+        this.set('value', true);
+        this.sendAction('onChange', {
+          checked: true
+        });
+      },
+      onUnchecked: () => {
+        this.set('value', false);
+        this.sendAction('onChange', {
+          checked: false
+        });
+      }
     });
-  }),
+  },
+
+  /**
+    Destroys DOM-related component's properties.
+  */
+  willDestroyElement() {
+    this._super(...arguments);
+
+    // Destroys Semantic UI checkbox.
+    this.$().checkbox('destroy');
+  },
+
+  /**
+    Component's action invoking when checkbox was clicked and it's 'checked' state changed.
+
+    @method sendingActions.onChange
+    @param {Object} e Action's event object.
+    @param {Boolean} e.checked Component's current value.
+  */
 });